Moved AppSettings to its own view
