+++ SongLab/Components/TrackingToolBar/TrackingToolbarView.swift
@@ -11,11 +11,13 @@ struct TrackingToolbarView: View {
     
     //MARK: - API
     
+    @Binding var isSettingsPresented: Bool
     @Binding var isRecording: Bool
     
-    init(audioManager: AudioManager, isRecording: Binding<Bool>) {
+    init(audioManager: AudioManager, isRecording: Binding<Bool>, isSettingsPresented: Binding<Bool>) {
         _viewModel = StateObject(wrappedValue: TrackingToolbarViewModel(audioManager: audioManager))
         _isRecording = isRecording
+        _isSettingsPresented = isSettingsPresented
     }
     
     //MARK: - Variables
@@ -32,7 +34,7 @@ struct TrackingToolbarView: View {
                 .padding(.leading)
             RecordButtonView(isRecording: $isRecording)
                 .padding(.horizontal, 25)
-            TrackingSettingsView()
+            appSettingsButton
                 .padding(.trailing)
         }
         .background(
@@ -51,11 +53,21 @@ struct TrackingToolbarView: View {
         .padding(.top, 25)
     }
     
+    var appSettingsButton: some View {
+        Button {
+            isSettingsPresented.toggle()
+        } label: {
+            Image(systemName: "slider.horizontal.3")
+        }
+        .foregroundStyle(.primary)
+        
+    }
+    
     var color: some View {
         colorScheme == .dark ? Color.black : Color.clear
     }
 }
 
 #Preview {
-    TrackingToolbarView(audioManager: MockAudioManager(), isRecording: .constant(false))
+    TrackingToolbarView(audioManager: MockAudioManager(), isRecording: .constant(false), isSettingsPresented: .constant(false))
 }
